inventory-service/pyproject.toml
[tool.poetry]
name = "inventory-service"
version = "0.1.0"
description = "Inventory management microservice"
authors = ["Milton Hernandez <miltonandresaxl@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.25"}
asyncpg = "^0.29.0"
alembic = "^1.13.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
aio-pika = "^9.3.0"
python-dotenv = "^1.0.0"
uuid = "^1.30"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.26.0"
ruff = "^0.1.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
